Moved creation of properties window
Property window was initially created (labels, layouts, etc...) in the create() function, which was called by mainwindow whenever the properties button was clicked on. This causes the property window to have to be remade every time the button is clicked. Moved the creation of the window to the constructor, that way it is only made once. Memory is not effected by this either, as the destructor isn't called until the parent window is closed, not when the child (property) window is closed.